+++ lib/hl7/message.ex
@@ -157,6 +157,10 @@ defmodule HL7.Message do
     msg
   end
 
+  def new(%HL7{} = msg, _options) do
+    HL7.to_list(msg) |> new() |> Map.put(:tag, msg.tags)
+  end
+
   def new(<<"MSH|^~\\&", _rest::binary>> = raw_text, options) do
     with {:ok, text} <- validate_text(raw_text, options) do
       copy = options[:copy] == true
@@ -226,6 +230,10 @@ defmodule HL7.Message do
 
   def to_list(%HL7.Message{segments: segments}) do
     segments
+  end
+
+  def to_list(%HL7.InvalidMessage{}) do
+    raise RuntimeError, "invalid HL7 data"
   end
 
   def to_list(%HL7.RawMessage{} = msg) do