+++ lib/hl7/path.ex
@@ -1,4 +1,9 @@
 defmodule HL7.Path do
+  @moduledoc """
+  A module and struct representing an HL7 path. This is typically created through the `HL7.sigil_p/2` macro
+  in order to gain compile-time correctness guarantees.
+  """
+
   defstruct segment: nil,
             segment_number: nil,
             field: nil,
@@ -9,11 +14,10 @@ defmodule HL7.Path do
             data: nil,
             path: nil
 
-  @type level() :: :segment | :field | :repetition | :component | :subcomponent
   @type t() :: %__MODULE__{}
 
   @doc ~S"""
-  Generates an `~p` sigil data structure at runtime.
+  Creates an HL7 path struct at runtime.
 
   ## Examples
 
@@ -95,6 +99,6 @@ end
 
 defimpl Inspect, for: HL7.Path do
   def inspect(%HL7.Path{path: path}, _opts) do
-    "~p[" <> path <> "]"
+    "~p\"" <> path <> "\""
   end
 end