+++ test/hl7/segment_test.exs
@@ -1,4 +1,4 @@
-defmodule HL7SegmentTest do
+defmodule HL7.SegmentTest do
   use ExUnit.Case
   require Logger